reorganize packages; add `tas.function.*`
