+++ src/main/java/tas/function/Tuple2.java
@@ -1,0 +1,11 @@
+package tas.function;
+
+public final class Tuple2<A,B> {
+  public final A _1;
+  public final B _2;
+
+  public Tuple2(A _1, B _2) {
+    this._1 = _1;
+    this._2 = _2;
+  }
+}
